pyproject.toml
[project]
name = "KGATE"
version = "0.2.15"
description = "Knowledge Graph Autoencoder Training Environment, bridging PyG encoders and TorchKGE decoders."
authors = [
    {name = "Benjamin Loire", email = "benjamin.loire@univ-amu.fr"},
    {name = "Galadriel Brière", email = "marie-galadriel.briere@univ-amu.fr"}
]
readme = "README.md"
license = "MIT"
keywords = ["Knowledge Graph", "Knowledge Graph Embedding", "Autoencoder", "Machine Learning","Link Prediction"]
classifiers = [
  # How mature is this project? Common values are
  #   3 - Alpha
  #   4 - Beta
  #   5 - Production/Stable
  "Development Status :: 3 - Alpha",

  # Indicate who your project is intended for
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

requires-python = ">=3.10"

[project.urls]
repository = "https://github.com/BAUDOTlab/KGATE"
issues = "https://github.com/BAUDOTlab/KGATE/issues"

[tool.poetry]
include = [
    { path = "config_template.toml", format = ["sdist","wheel"] }
]

[tool.poetry.dependencies]
python = ">=3.10"
torchkge = "^0.17.7"
torch-geometric = "^2.6.1"
matplotlib = "^3.10.0"
pytorch-ignite = "^0.5.1"
pyyaml = "^6.0.2"
pytest = "^8.3.4"
optuna = "^4.2.1"
tomli-w = "^1.2.0"
